Committing a more full-featured version of Porthos
In contrast to porthos-basic, porthos.py is meant to be useful, rather than
serving as a proof of concept and potential guide for other programmers.
This version is not restricted to native Python modules, and the readibility
of the code has only been improved so far as it's helped me debug it, which is
not to imply that I have thoroughly tested this program yet. At this point,
the output merely "looks" correct based on a few simply test files, so caveat
emptor for now.

Signed-off-by: Ted Gibbons <ted.codes@gmail.com>